LayoutTests/imported/w3c/web-platform-tests/dom/events/EventListener-incumbent-global-2.sub-expected.txt



FAIL Check the incumbent global EventListeners  are called with assert_equals: Incumbent should have been the caller of addEventListener() expected "http://www1.web-platform.test:8800" but got "http://www2.web-platform.test:8800"
